+++ packages/powersync/lib/src/streaming_sync.dart
@@ -198,6 +198,7 @@ class StreamingSyncImplementation {
   /// To clear errors, use [_noError] instead of null.
   void _updateStatus(
       {DateTime? lastSyncedAt,
+      bool? hasSynced,
       bool? connected,
       bool? connecting,
       bool? downloading,
@@ -209,6 +210,7 @@ class StreamingSyncImplementation {
         connected: c,
         connecting: !c && (connecting ?? lastStatus.connecting),
         lastSyncedAt: lastSyncedAt ?? lastStatus.lastSyncedAt,
+        hasSynced: hasSynced ?? lastStatus.hasSynced,
         downloading: downloading ?? lastStatus.downloading,
         uploading: uploading ?? lastStatus.uploading,
         uploadError: uploadError == _noError